Bluetooth: Do not repeatedly ask for authentication
Removing code, not present in upstream, that causes kernel to
repeatedly ask for authentication when the link key is invalid.

If two devices were paired but the remote device had removed
the pairing, a connection attempt would result in authentication
failure. The removed code would in this case ask for authentication
again, the remote device would reply with a link key request that
would be handled by the userspace, and when the (still invalid)
link key was received authentication would fail again and the
process would unnecessarily repeat.

Removing the code stops the unnecessary repeated authentication
attempts with an invalid link key.

[sbj] Do not repeatedly ask for authentication. Fixes JB#18165.

Signed-off-by: Hannu Mallat <hannu.mallat@jolla.com>